requirements.txt
google-genai
feedparser
requests
python-dateutil
pytz